requirements.txt
numpy
protobuf==3.20.*
pillow
matplotlib
opencv-python-headless
tensorflow
streamlit-drawable-canvas
pyarrow